style: terminal - change font for a bolder bold that's actually noticeable
JetBrains Mono is fantastic, but it's bold is not very bold and I couldn't get Wezterm to use a bolder bold (although this may be worth spending more time on in the future.)

Switching to Fira Mono for now. Also added a helpful `styles.txt`, courteosy yet again of Wez, that can be `cat`ed to quickly test terminal font capabilities.